Room and Dungeon updates for ToString
