+++ DungeonTest.java
@@ -1,0 +1,10 @@
+import dungeon.*;
+public class DungeonTest {
+   public static void main(String[] args) {
+      Dungeon dungeon = new Dungeon(5);
+      System.out.print(dungeon.toString());
+   
+   
+   }
+
+}
