--- Skeleton.java
@@ -1,14 +1,0 @@
-public class Skeleton extends Monster {
-
-   public Skeleton() {
-      super("Skeletor the Skeleton Warrior", Stats.getSkeletonStats());
-   }
-   
-   @Override
-   public void attack(DungeonCharacter opponent) {
-      System.out.println(super.getName() + " slices his rusty longsword at " + opponent.getName() + ":");
-      super.attack(opponent);
-   }
-}
-
-
