+++ dungeon/Room.java
@@ -1,110 +1,75 @@
+package dungeon;
 public class Room {
 
     private boolean hasExit = false;
     private boolean hasEntrance = false;
     private boolean hasMonster = false;
-    private boolean hasHealthPotion = false;
+    private boolean hasHealthpotion = false;
     private boolean hasVisionPotion = false;
-    private boolean hasPillar = false;
+    private boolean hasPiller = false;
+    private boolean visited = true;
     private int[] coordinates;
 
-    public Room(final boolean hasExit, final boolean hasEntrance, final boolean hasMonster, final boolean hasHealthPotion, final boolean hasVisionPotion, final boolean hasPillar, final int[] coordinates) {
-      setHasExit(hasExit);
-      setHasEntrance(hasEntrance);
-      setHasMonster(hasMonster);
-      setHasHealthPotion(hasHealthPotion);
-      setHasVisionPotion(hasVisionPotion);
-      setHasPillar(hasPillar);
-      setCoordinates(coordinates);
+    public Room(boolean hasMonster,boolean hasHealthPotion , boolean hasVisionPotion, int[] coordinates){
+        setHasExit(false);
+        setHasEntrance(false);
+        setHasMonster(hasMonster);
+        setHasHealthPotion(hasHealthPotion);
+        setHasVisionPotion(hasVisionPotion);
+        setHasPillar(false);
+        setCoordinates(coordinates);
+    }
+
+    public boolean getHasExit() {
+      return hasExit;
+    }
+    
+    public boolean getVisited() {
+      return visited;    
+    }
+    
+    public void isVisited() {
+      this.visited = true;
     }
     
     private char whatIsInRoom() {
-      boolean[] arr = {isHasExit(), isHasEntrance(), isHasMonster(), isHasHealthPotion(), isHasVisionPotion(), isHasPillar()};
-      int numOfTrue = 0;
-      for(int i = 0; i < arr.length; i++) {
-         if(arr[i])
-            numOfTrue++;
-      }
-      
-      if(numOfTrue == 1)
-         if(isHasExit())
-            return 'O';
-         else if(isHasEntrance())
-            return 'I';
-         else if(isHasMonster())
-            return 'X';
-         else if(isHasHealthPotion())
-            return 'H';
-         else if(isHasVisionPotion())
-            return 'V';
-         else if(isHasPillar())
-            return 'C';
-      
-      else if(numOfTrue > 1)
-         return 'M';
-      else
-         return ' ';
-      
-      return ' ';
-    }
-    
-    @Override
-    public String toString() {
-      char c = whatIsInRoom();
-      
-      //Needs work when drawing out the room info.
-      if(this.coordinates[1] == 0) { //left side
-         if(this.coordinates[0] == 0) 
-            return "* * *\n* "+c+" |\n* - *"; //upper corner                
-         else if(this.coordinates[0] == 4) //lower corner
-            return "* - *\n* "+c+" |\n* * *";
-         else
-            return "* - *\n* "+c+" |\n* - *"; //West walls
-      }
-      
-      else if(this.coordinates[1] == 4) { //right side
-         if(this.coordinates[0] == 0) 
-            return "* * *\n| "+c+" * \n* - *"; //upper corner
-         else if(this.coordinates[0] == 4) //lower corner
-            return "* - *\n| "+c+" *\n* * *";
-         else
-            return "* - *\n| "+c+" *\n* - *"; //East Wall
-      }
-      
-      else {
-         if(this.coordinates[0] == 0) //North Wall
-            return " * *\n"+c+" |\n - *"; 
-         else if(this.coordinates[0] == 4) //South Wall
-            return " - *\n"+c+" |\n * *";  
-         else 
-            return " -  \n "+c+" \n  - ";
-      } 
-    }
+        boolean[] arr = {getHasExit(), getHasEntrance(), getHasMonster(), getHasHealthPotion(), getHasVisionPotion(), getHasPillar()};
+        int numOfTrue = 0;
+        for (int i = 0; i < arr.length; i++) {
+            if (arr[i])
+                numOfTrue++;
+        }
+        if (numOfTrue == 1){
+            if (getHasExit())
+                return 'O';
+            else if (getHasEntrance())
+                return 'I';
+            else if (getHasMonster())
+                return 'X';
+            else if (getHasHealthPotion())
+                return 'H';
+            else if (getHasVisionPotion())
+                return 'V';
+            else if (getHasPillar())
+                return 'C';
+        }else if(numOfTrue > 1) {
+            return 'M';
+        }
 
-    public void setCoordinates(final int[] coordinates) {
-      this.coordinates = coordinates;
+        return ' ';
     }
-    
-    public int[] getCoordinates() {
-      return this.coordinates;
-    }
-    
-    public boolean isHasExit() {
-        return hasExit;
-    }
-
     public void setHasExit(boolean hasExit) {
         this.hasExit = hasExit;
     }
 
-    public boolean isHasEntrance() {
+    public boolean getHasEntrance() {
         return hasEntrance;
     }
 
     public void setHasEntrance(boolean hasEntrance) {
         this.hasEntrance = hasEntrance;
     }
-    public boolean isHasMonster() {
+    public boolean getHasMonster() {
         return hasMonster;
     }
 
@@ -112,15 +77,15 @@ public class Room {
         this.hasMonster = hasMonster;
     }
 
-    public boolean isHasHealthPotion() {
-        return hasHealthPotion;
+    public boolean getHasHealthPotion() {
+        return hasHealthpotion;
     }
 
-    public void setHasHealthPotion(boolean hasHealthPotion) {
-        this.hasHealthPotion = hasHealthPotion;
+    public void setHasHealthPotion(boolean hasHealthpotion) {
+        this.hasHealthpotion = hasHealthpotion;
     }
 
-    public boolean isHasVisionPotion() {
+    public boolean getHasVisionPotion() {
         return hasVisionPotion;
     }
 
@@ -128,12 +93,53 @@ public class Room {
         this.hasVisionPotion = hasVisionPotion;
     }
 
-    public boolean isHasPillar() {
-        return hasPillar;
+    public boolean getHasPillar() {
+        return hasPiller;
     }
 
     public void setHasPillar(boolean hasPillar) {
-        this.hasPillar = hasPillar;
+        this.hasPiller = hasPillar;
     }
 
+    public void setCoordinates(final int[] coordinates) {
+        this.coordinates = coordinates;
+    }
+
+    public int[] getCoordinates() {
+        return this.coordinates;
+    }
+
+    @Override
+    public String toString() {
+        isVisited(); //can only be drawn if it was visited.
+        char c = whatIsInRoom();
+
+        //Needs work when drawing out the room info.
+        if(this.coordinates[1] == 0) { //left side
+            if(this.coordinates[0] == 0)
+                return "* * *\n* "+c+" |\n* - *"; //upper corner
+            else if(this.coordinates[0] == 4) //lower corner
+                return "* - *\n* "+c+" |\n* * *";
+            else
+                return "* - *\n* "+c+" |\n* - *"; //West walls
+        }
+
+        else if(this.coordinates[1] == 4) { //right side
+            if(this.coordinates[0] == 0)
+                return "* * *\n| "+c+" * \n* - *"; //upper corner
+            else if(this.coordinates[0] == 4) //lower corner
+                return "* - *\n| "+c+" *\n* * *";
+            else
+                return "* - *\n| "+c+" *\n* - *"; //East Wall
+        }
+
+        else {
+            if(this.coordinates[0] == 0) //North Wall
+                return " * *\n"+c+" |\n - *";
+            else if(this.coordinates[0] == 4) //South Wall
+                return " - *\n"+c+" |\n * *";
+            else
+                return " -  \n "+c+" \n  - ";
+        }
+    }
 }
